moved linking to app root as suggested
