+++ screens/more/reports.js
@@ -12,7 +12,7 @@ const ReportsScreen = (props) => {
          */
       }
       {
-        //<CalorieReport />
+        <CalorieReport />
       }
     {
         /* piechart
